% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-module.R
\name{alignmentModule}
\alias{alignmentModule}
\title{Server part of alignment module}
\usage{
alignmentModule(input, output, session,
  datatable_opts = pavian::datatable_opts)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session}

\item{datatable_opts}{Additional options for creating the datatable.}
}
\value{
Alignment module server functionality
}
\description{
Server part of alignment module
}
