% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_alignment.R
\name{alignmentModuleUI}
\alias{alignmentModuleUI}
\title{UI part of alignment module}
\usage{
alignmentModuleUI(id)
}
\arguments{
\item{id}{Shiny namespace id}
}
\value{
UI part of alignment module
}
\description{
UI part of alignment module
}
