% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{beautify_colnames}
\alias{beautify_colnames}
\title{Beautify colnames}
\usage{
beautify_colnames(x)
}
\arguments{
\item{x}{data.frame or matrix}
}
\value{
data.frame or matrix with nicer colnames
}
\description{
Beautify colnames
}
