% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{beautify_string}
\alias{beautify_string}
\title{Upper-case first letter of string and use white-space for dot and underscore}
\usage{
beautify_string(x)
}
\arguments{
\item{x}{string}
}
\value{
beautified string
}
\description{
Upper-case first letter of string and use white-space for dot and underscore
}
\examples{
beatuify_string(c("this.is.not.beautiful"))
}
