% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_comparision.R
\name{comparisionModule}
\alias{comparisionModule}
\title{Title}
\usage{
comparisionModule(input, output, session, samples_df, reports,
  datatable_opts = NULL)
}
\arguments{
\item{pattern}{}
}
\description{
Title
}

