% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_comparison2.R
\name{comparisonModuleUI}
\alias{comparisonModuleUI}
\title{UI part of the comparison module}
\usage{
comparisonModuleUI(id)
}
\arguments{
\item{id}{Shiny namespace id.}
}
\value{
UI elements of the comparison module.
}
\description{
UI part of the comparison module
}
