% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp-ui.R
\name{dashboardUI}
\alias{dashboardUI}
\title{Function building dashboard UI, used in Shiny app}
\usage{
dashboardUI(request)
}
\arguments{
\item{request}{request object}
}
\value{
Dashboard page
}
\description{
Function building dashboard UI, used in Shiny app
}
