% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_datainput.R
\name{dataInputModuleUI}
\alias{dataInputModuleUI}
\title{UI part of pavian data input module}
\usage{
dataInputModuleUI(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
shiny UI elements
}
\description{
UI part of pavian data input module
}

