% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison-merge_reports.R
\name{filter_cladeReads}
\alias{filter_cladeReads}
\title{Remove specified taxa or whole clades this function removes the read numbers from the parents, too}
\usage{
filter_cladeReads(cladeReads, tax_data, rm_taxa)
}
\arguments{
\item{cladeReads}{clade "cladeReads"}

\item{tax_data}{tax data}

\item{rm_taxa}{taxa to remove}
}
\value{
filtered clade "cladeReads"
}
\description{
Remove specified taxa or whole clades this function removes the read numbers from the parents, too
}
