% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison-merge_reports.R
\name{filter_taxa}
\alias{filter_taxa}
\title{Filter taxon}
\usage{
filter_taxa(tax_data, rm_clades = NULL, rm_taxa = NULL,
  taxRank = "-")
}
\arguments{
\item{tax_data}{tax_data}

\item{rm_clades}{clades to remove}

\item{rm_taxa}{taxa to remove}

\item{taxRank}{taxRank to keep}
}
\value{
selected rows
}
\description{
Filter taxon
}
