% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_krakenres.R
\name{filter_taxon}
\alias{filter_taxon}
\title{Filter lines from a kraken report result based on the taxonomy name}
\usage{
filter_taxon(krakenres, filter_taxon, do_message = TRUE)
}
\arguments{
\item{krakenres}{kraken report.}

\item{filter_taxon}{name of entry to remove.}

\item{do_message}{If TRUE, report how many rows and reads were deleted.}
}
\value{
filtered krakenres
}
\description{
It updates the read_stay counts, and removes any children below the
entry, and any parent entries that have no reads that stay
}
\examples{
\\donotrun{
  filter_taxon(krakenres, 's_Homo sapiens')
}
}

