% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltools-getDependency.R
\name{getDependency}
\alias{getDependency}
\title{htmlwidgets:::getDependency}
\usage{
getDependency(name, package = name)
}
\arguments{
\item{name}{name}

\item{package}{package}
}
\value{
list of dependencies
}
\description{
Copied to circumvent R CMD check warning when using non-exported function from a package
}
\examples{
getDependency('sparkline')
}

