% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-humanreadable.R
\name{humanreadable}
\alias{humanreadable}
\title{Make a number human readable}
\usage{
humanreadable(x, base = 10^3, suffix = "", signif_digits = 3)
}
\arguments{
\item{x}{Number.}

\item{base}{typically 1000 or 2^10.}

\item{suffix}{Will be concatendated to the end.}

\item{signif_digits}{Number of significant digits.}
}
\value{
human readable representation of the number
}
\description{
Make a number human readable
}
\examples{
humanreadable(1234.567)
}
