% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kraken_sunburst.R
\name{kraken_sunburst}
\alias{kraken_sunburst}
\title{kraken_sunburst: Create a}
\usage{
kraken_sunburst(krakenres)
}
\arguments{
\item{krakenres}{kraken report, read with read.krakenres}
}
\value{
sunburst widget
}
\description{
kraken_sunburst: Create a
}

