% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.or.create.R
\name{load.or.create}
\alias{load.or.create}
\title{Load a file from a cache directory, or create and save it}
\usage{
load.or.create(f, name, recreate = FALSE, cache_dir = "cache")
}
\arguments{
\item{f}{function to call when the file does not exist.}

\item{name}{name of cache file (including .rds).}

\item{recreate}{if TRUE, recreate the file even if it exists.}

\item{cache_dir}{directory that contains cache files. It is created if it does not exist.}
}
\value{
object created by function f, or stored in cache file
}
\description{
Load a file from a cache directory, or create and save it
}
\examples{
\\donotrun{
load.or.create(function() {
  rnorm(10000)
}, "rnorm-results.rds", recreate=FALSE, cache_dir="cache")
}
}

