% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-load_or_create.R
\name{load_or_create}
\alias{load_or_create}
\title{Load a file from a cache directory, or create and save it}
\usage{
load_or_create(f, name, recreate = FALSE, cache_dir = "cache")
}
\arguments{
\item{f}{function to call when the file does not exist.}

\item{name}{name of cache file (including .rds).}

\item{recreate}{if TRUE, recreate the file even if it exists.}

\item{cache_dir}{directory that contains cache files. It is created if it does not exist.}
}
\value{
object created by function f, or stored in cache file
}
\description{
Load a file from a cache directory, or create and save it
}
\examples{
\dontrun{
 load_or_create(function() {
   rnorm(10000)
 }, "rnorm-results.rds", recreate=FALSE, cache_dir="cache")
}
}
