% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison-merge_reports.R
\name{normalize}
\alias{normalize}
\title{Helper function to normalize data}
\usage{
normalize(m, sums = colSums(m, na.rm = T))
}
\arguments{
\item{m}{matrix}

\item{sums}{column sums}
}
\value{
normalized matrix
}
\description{
Helper function to normalize data
}
