% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/padded_sort.R
\name{padded_sort}
\alias{padded_sort}
\title{Sort a vector of string, but pad the first numbers with zeros before doing so.}
\usage{
padded_sort(x, n_pad = 5)
}
\arguments{
\item{x}{vector of strings}

\item{n_pad}{number of zeros the numbers are padded with}
}
\value{
vector of strings sorted 'alpha-numerically'
}
\description{
Adapted from http://stackoverflow.com/a/15552536
}
\examples{
strings <- c("A1","A01","A10","A20","A")
paddedsort(strings)
}

