% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paddedsort.R
\name{paddedsort}
\alias{paddedsort}
\title{Sort a vector of string, but padd the numbers with zeros before doing so.}
\usage{
paddedsort(x, n.pad = 5)
}
\arguments{
\item{x}{vector of strings}

\item{pad}{number of zeros the numbers are padded with}
}
\value{
vector of strings sorted 'alpha-numerically'
}
\description{
Adapted from http://stackoverflow.com/a/15552536
}
\examples{
strings <- c("A1","A01","A10","A20")
paddedsort(strings)
}

