% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp-server.R
\name{pavianServer}
\alias{pavianServer}
\title{Pavian server function}
\usage{
pavianServer(input, output, session)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Session object}
}
\description{
Pavian server function
}
