% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_level_reads.R
\name{plot_level_reads}
\alias{plot_level_reads}
\title{Barplot of metagenomics results at a certain level}
\usage{
plot_level_reads(level_reads, value = "perc", ylab = "fraction of reads",
  labels = percent)
}
\arguments{
\item{level_reads}{\code{data.frame} with all reads at a certain level with columns '.id', 'name', value}

\item{value}{Name of column with quantitative values to display}

\item{ylab}{Label on y axis}

\item{labels}{Label formatter. See \code{\link[ggplot2]{scale_y_continuous}}.}
}
\value{
Barplot of level reads
}
\description{
Barplot of metagenomics results at a certain level
}

