% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pileup.R
\name{plot_pileup}
\alias{plot_pileup}
\title{Title}
\usage{
plot_pileup(pileup, nreads, seq_lengths, align_loess, show_step = TRUE,
  nwin = 1000, text_size = 4)
}
\arguments{
\item{nwin}{}
}
\description{
Title
}

