% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_level_reads.R
\name{plot_taxRank_cladeReads}
\alias{plot_taxRank_cladeReads}
\title{Barplot of metagenomics results at a certain taxRank}
\usage{
plot_taxRank_cladeReads(taxRank_cladeReads, value = "perc",
  ylab = "fraction of reads", labels = percent)
}
\arguments{
\item{taxRank_cladeReads}{\code{data.frame} with all cladeReads at a certain taxRank with columns '.id', 'name', value}

\item{value}{Name of column with quantitative values to display}

\item{ylab}{Label on y axis}

\item{labels}{Label formatter. See \code{\link[ggplot2]{scale_y_continuous}}.}
}
\value{
Barplot of taxRank "cladeReads"
}
\description{
Barplot of metagenomics results at a certain taxRank
}
