% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.krakenres.R
\name{read.krakenres}
\alias{read.krakenres}
\title{Read kraken report}
\usage{
read.krakenres(myfile, collapse = TRUE, keep.levels = c("D", "K", "P", "C",
  "O", "F", "G", "S"), min.depth = 0, filter_taxon = NULL,
  has.header = NULL, add.level.columns = FALSE)
}
\arguments{
\item{myfile}{kraken report file}

\item{collapse}{should the results be collapsed to only those levels specified in keep.levels?}

\item{keep.levels}{levels to keep when collapse is TRUE}

\item{min.depth}{minimum depth}

\item{filter_taxon}{filter certain taxon names}

\item{add.level.columns}{if TRUE, for each level columns are added}

\item{v2}{report format is in v2, with kmer frequencies}
}
\value{
report data.frame
}
\description{
Read kraken report
}

