% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_krakenres.R
\name{read_krakenres}
\alias{read_krakenres}
\title{Read kraken report}
\usage{
read_krakenres(myfile, collapse = TRUE, keep_levels = c("D", "K", "P", "C",
  "O", "F", "G", "S"), min.depth = 0, filter_taxon = NULL,
  has_header = NULL, add_level_columns = FALSE)
}
\arguments{
\item{myfile}{kraken report file}

\item{collapse}{should the results be collapsed to only those levels specified in keep_levels?}

\item{keep_levels}{levels to keep when collapse is TRUE}

\item{min.depth}{minimum depth}

\item{filter_taxon}{filter certain taxon names}

\item{has_header}{if the kraken report has a header or not}

\item{add_level_columns}{if TRUE, for each level columns are added}
}
\value{
report data.frame
}
\description{
Read kraken report
}

