% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_krakenres.R
\name{read_report}
\alias{read_report}
\title{Read kraken and centrifuge-style reports}
\usage{
read_report(myfile, has_header = NULL)
}
\arguments{
\item{myfile}{report file}

\item{has_header}{if the kraken report has a header or not}
}
\value{
report data.frame
}
\description{
Read kraken and centrifuge-style reports
}

