% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_report.R
\name{read_report2}
\alias{read_report2}
\title{Read kraken or centrifuge-style report}
\usage{
read_report2(myfile, collapse = TRUE, keep_ranks = c("D", "K", "P", "C",
  "O", "F", "G", "S"), min.depth = 0, filter_taxon = NULL,
  has_header = NULL, add_rank_columns = FALSE)
}
\arguments{
\item{myfile}{kraken report file}

\item{collapse}{should the results be collapsed to only those ranks specified in keep_ranks?}

\item{keep_ranks}{ranks to keep when collapse is TRUE}

\item{min.depth}{minimum depth}

\item{filter_taxon}{filter certain taxon names}

\item{has_header}{if the kraken report has a header or not}

\item{add_rank_columns}{if TRUE, for each rank columns are added}
}
\value{
report data.frame
}
\description{
Read kraken or centrifuge-style report
}
