% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sample_data.R
\name{read_sample_data}
\alias{read_sample_data}
\title{Read sample_data.csv from directory, or create it}
\usage{
read_sample_data(my_dir, def_filename = "sample_data.csv", ext = c("report",
  "profile"))
}
\arguments{
\item{my_dir}{directory}

\item{def_filename}{filename}

\item{ext}{extension of report files}
}
\value{
sample_data data.frame
}
\description{
Read sample_data.csv from directory, or create it
}
