% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_server_directory.R
\name{read_server_directory1}
\alias{read_server_directory1}
\title{Title}
\usage{
read_server_directory1(data_dir, sample_set_name = NULL,
  include_base_dir = T, display_messages = TRUE)
}
\arguments{
\item{data_dir}{data directory}

\item{sample_set_name}{name of sample set}

\item{include_base_dir}{include base directory}

\item{display_messages}{display messages?}
}
\value{
resulting sample sets
}
\description{
Title
}
