% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_overview_table.R
\name{reportOverviewModule}
\alias{reportOverviewModule}
\title{Shiny modules to display an overview of metagenomics reports}
\usage{
reportOverviewModule(input, output, session, samples_df, reports,
  datatable_opts = NULL)
}
\arguments{
\item{session}{}
}
\description{
Shiny modules to display an overview of metagenomics reports
}

