% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview-module.R
\name{reportOverviewModuleUI}
\alias{reportOverviewModuleUI}
\title{UI part of report overview module}
\usage{
reportOverviewModuleUI(id)
}
\arguments{
\item{id}{Shiny namespace id.}
}
\value{
UI elements for report overview module.
}
\description{
UI part of report overview module
}
