% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison-merge_reports.R
\name{robust_zscore}
\alias{robust_zscore}
\title{Calculate robust z-score on matrix}
\usage{
robust_zscore(m, min_scale = 1)
}
\arguments{
\item{m}{matrix}

\item{min_scale}{minimum scale}
}
\value{
robust z-score of matrix
}
\description{
Calculate robust z-score on matrix
}
