% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runCentrifuger.R
\name{runCentrifuger}
\alias{runCentrifuger}
\title{Run Centrifuger web interface}
\usage{
runCentrifuger(cache_dir = "cache")
}
\arguments{
\item{cache_dir}{Directory to save temporary files}
}
\description{
Run Centrifuger web interface
}

