% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_sample.R
\name{sampleModule}
\alias{sampleModule}
\title{Server part for sample module}
\usage{
sampleModule(input, output, session, sample_data, reports,
  datatable_opts = NULL)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{sample_data}{Samples \code{data.frame}}

\item{reports}{List of reports}

\item{datatable_opts}{Additional options for datatable}
}
\value{
Sample module server functionality
}
\description{
Server part for sample module
}

