% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-module.R
\name{sampleModuleUI}
\alias{sampleModuleUI}
\title{UI part for sample module}
\usage{
sampleModuleUI(id)
}
\arguments{
\item{id}{Shiny namespace id}
}
\value{
UI elements for sample module
}
\description{
UI part for sample module
}
