% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarize.kraken.report.R
\name{summarize.kraken.report}
\alias{summarize.kraken.report}
\title{Summarize the result of a kraken report}
\usage{
summarize.kraken.report(my.report)
}
\arguments{
\item{my.report}{kraken report}
}
\value{
data.frame with number of raw read, classified at certain levels, etc
}
\description{
Summarize the result of a kraken report
}
\examples{
\\donotrun{
  summarize.kraken.report(krakenres)
}
}

