% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarize_kraken_report.R
\name{summarize_kraken_report}
\alias{summarize_kraken_report}
\title{Summarize the result of a kraken report}
\usage{
summarize_kraken_report(my_report)
}
\arguments{
\item{my_report}{kraken report}
}
\value{
data.frame with number of raw read, classified at certain levels, etc
}
\description{
Summarize the result of a kraken report
}
\examples{
\\donotrun{
  summarize_kraken_report(krakenres)
}
}

