% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview-summarize_report.R
\name{summarize_reports}
\alias{summarize_reports}
\title{Summarize list of reports}
\usage{
summarize_reports(reports)
}
\arguments{
\item{reports}{list of reports}
}
\value{
summary of list of reports
}
\description{
Summarize list of reports
}
