% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{zero_if_na}
\alias{zero_if_na}
\title{Helper function that sets NAs to zeros in a supplied data.frame}
\usage{
zero_if_na(df)
}
\arguments{
\item{df}{data.frame or matrix}
}
\value{
data.frame or matrix in which all negative and NA values are set to zero
}
\description{
Helper function that sets NAs to zeros in a supplied data.frame
}
