% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_alignment.R
\name{alignmentModule}
\alias{alignmentModule}
\title{Server part of alignment module}
\usage{
alignmentModule(input, output, session, samples_df)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session}

\item{samples_df}{\code{data.frame} for samples}
}
\value{
Alignment module server functionality
}
\description{
Server part of alignment module
}

