% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_comparison.R
\name{comparisonModule}
\alias{comparisonModule}
\title{Title}
\usage{
comparisonModule(input, output, session, samples_df, reports,
  datatable_opts = NULL, filter_func = NULL)
}
\arguments{
\item{pattern}{}
}
\description{
Title
}

