% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_datainput.R
\name{dataInputModule}
\alias{dataInputModule}
\title{Title}
\usage{
dataInputModule(input, output, session, ..., pattern = "defs.csv$",
  cache_tree = TRUE)
}
\arguments{
\item{pattern}{}
}
\description{
Title
}

