% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_datainput.R
\name{dataInputModuleUI}
\alias{dataInputModuleUI}
\title{UI part of pavian data input module}
\usage{
dataInputModuleUI(id, server_access = TRUE,
  start_with = getOption("pavian.start_data_input_with", "Example data"))
}
\arguments{
\item{id}{Namespace ID.}

\item{server_access}{Whether to allow access to server directories.}

\item{start_with}{One of 'example_data', 'upload', and 'server' (if server_access is TRUE).}
}
\value{
Shiny UI elements
}
\description{
UI part of pavian data input module
}

