% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datainput-module.R
\name{dataInputModuleUI}
\alias{dataInputModuleUI}
\title{UI part of pavian data input module}
\usage{
dataInputModuleUI(id, server_access = getOption("pavian.server_access",
  default = FALSE),
  start_with = getOption("pavian.start_data_input_with", "Upload files"))
}
\arguments{
\item{id}{Namespace ID.}

\item{server_access}{Whether to allow access to server directories.}

\item{start_with}{One of 'Example data', 'Upload files', and 'Use data on server' (if server_access is TRUE).}
}
\value{
Shiny UI elements
}
\description{
UI part of pavian data input module
}
