% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datainput-read_report.R
\name{filter_taxon}
\alias{filter_taxon}
\title{Filter lines from a kraken report result based on the taxonomy name}
\usage{
filter_taxon(report, filter_taxon, rm_clade = TRUE, do_message = FALSE)
}
\arguments{
\item{report}{Report \code{data.frame}.}

\item{filter_taxon}{Name of entry to remove.}

\item{rm_clade}{If \code{TRUE}, remove all cladeReads at and below clade, otherwise just set the number of cladeReads that stay at taxon to zero.}

\item{do_message}{If \code{TRUE}, report how many rows and cladeReads were deleted.}
}
\value{
filtered report
}
\description{
It updates the read_stay counts, and removes any children below the
entry, and any parent entries that have no "cladeReads" that stay
}
