% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_level_reads.R
\name{get_level_reads}
\alias{get_level_reads}
\title{Get all reads from a certain level}
\usage{
get_level_reads(reports, ..., min.perc = 0.1, take.names = NULL)
}
\arguments{
\item{reports}{list fo kraken reports}

\item{...}{subsetting arguments}

\item{min.perc}{minimum percentage to be included}
}
\value{
all reads from a certain level
}
\description{
Get all reads from a certain level
}
\examples{
plot_level_reads(get_level_reads(reports,level=="P",min.perc=.05)) + ggtitle("Phylum level")
plot_level_reads(get_level_reads(reports,level=="C",min.perc=.05)) + ggtitle("Clade level")
plot_level_reads(get_level_reads(reports,level=="O",min.perc=.05)) + ggtitle("Order level")
plot_level_reads(get_level_reads(reports,level=="F" & name != "f_Hominidae",min.perc=.05)) + ggtitle("Family level (w/o human)")
}

