% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-get_pileup.R
\name{get_nreads}
\alias{get_nreads}
\title{Get number of aligned reads from a bam file}
\usage{
get_nreads(bam_file, min_mapq = 0)
}
\arguments{
\item{bam_file}{character(1) or BamFile; BAM file path.}

\item{min_mapq}{Minimum mapping quality}
}
\value{
Number of reads for each sequence.
}
\description{
Get number of aligned reads from a bam file
}
