% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-get_pileup.R
\name{get_seqlengths}
\alias{get_seqlengths}
\title{Get sequence lengths from a bam file}
\usage{
get_seqlengths(bam_file)
}
\arguments{
\item{bam_file}{character(1) or BamFile; BAM file path.}
}
\value{
named vector with lengths of each sequence.
}
\description{
Get sequence lengths from a bam file
}
