% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison-merge_reports.R
\name{merge_reports}
\alias{merge_reports}
\title{Merge report files into a wide format, with column(s) for each
report.}
\usage{
merge_reports(my_reports, numeric_col = c("cladeReads", "taxonReads"))
}
\arguments{
\item{my_reports}{Report data.frames.}

\item{numeric_col}{Numeric columns to keep.}
}
\value{
Combined data.frame
}
\description{
Merge report files into a wide format, with column(s) for each
report.
}
