% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_report.R
\name{merge_reports2}
\alias{merge_reports2}
\title{Merge report files into a wide format, with column(s) for each
report.}
\usage{
merge_reports2(my_reports, col_names = NULL)
}
\arguments{
\item{my_reports}{Report data.frames.}

\item{col_names}{Column names.}
}
\value{
Combined data.frame
}
\description{
Merge report files into a wide format, with column(s) for each
report.
}
