% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison-merge_reports.R
\name{merge_reports2}
\alias{merge_reports2}
\title{Merge report files into a wide format, with column(s) for each
report.}
\usage{
merge_reports2(my_reports, col_names = NULL, fix_taxnames = TRUE,
  update_progress = FALSE, id_cols = c("name", "taxRank", "taxID",
  "taxLineage"), numeric_cols = c("cladeReads", "taxonReads"))
}
\arguments{
\item{my_reports}{Report data.frames.}

\item{col_names}{Column names.}

\item{fix_taxnames}{Check if there are differences in the taxonomies of the reports.}

\item{update_progress}{Update shiny progress}

\item{id_cols}{Columns for taxdata}

\item{numeric_cols}{Numeric data}
}
\value{
Combined data.frame
}
\description{
Merge report files into a wide format, with column(s) for each
report.
}
\examples{
report_dir <- system.file("shinyapp", "example-data", "brain-biopsies", package="pavian")
reports <- read_reports(report_dir)
}
