% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_level_reads.R
\name{plot_level_reads}
\alias{plot_level_reads}
\title{Plot report content in form of a barplot}
\usage{
plot_level_reads(level_reads, value = "perc", ylab = "fraction of reads",
  labels = percent)
}
\arguments{
\item{labels}{}
}
\value{
barplot
}
\description{
Plot report content in form of a barplot
}

