% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-plot_pileup.R
\name{plot_pileup}
\alias{plot_pileup}
\title{Plot a pileup}
\usage{
plot_pileup(pileup, seq_info_df = NULL, nwin = 500, text_size = 5)
}
\arguments{
\item{pileup}{The output from \code{\link{get_pileup}}}

\item{seq_info_df}{Data.frame with information on sequences}

\item{nwin}{Number of points to evaluated to evaluate smoother at, when \code{show_loess} is \code{TRUE}}

\item{text_size}{Legend text size}
}
\value{
ggplot of pileup
}
\description{
Plot a pileup
}
