% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datainput-read_report.R
\name{read_report}
\alias{read_report}
\title{Read Kraken-style and MetaPhlAn reports}
\usage{
read_report(myfile, has_header = NULL, check_file = FALSE)
}
\arguments{
\item{myfile}{Kraken-style or MetaPhlAn report file.}

\item{has_header}{If the kraken report has a header or not.}

\item{check_file}{If TRUE, only the first 5 lines of the file are loaded.}
}
\value{
report data.frame
}
\description{
Read Kraken-style and MetaPhlAn reports
}
