% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datainput-read_report.R
\name{read_report2}
\alias{read_report2}
\title{Read kraken or centrifuge-style report}
\usage{
read_report2(myfile, collapse = TRUE, keep_taxRanks = c("D", "K", "P",
  "C", "O", "F", "G", "S"), min.depth = 0, filter_taxon = NULL,
  has_header = NULL, add_taxRank_columns = FALSE)
}
\arguments{
\item{myfile}{kraken report file}

\item{collapse}{should the results be collapsed to only those taxRanks specified in keep_taxRanks?}

\item{keep_taxRanks}{taxRanks to keep when collapse is TRUE}

\item{min.depth}{minimum depth}

\item{filter_taxon}{filter certain taxon names}

\item{has_header}{if the kraken report has a header or not}

\item{add_taxRank_columns}{if TRUE, for each taxRank columns are added}
}
\value{
report data.frame
}
\description{
Read kraken or centrifuge-style report
}
