% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_reports.R
\name{read_reports}
\alias{read_reports}
\title{Read report files, with Shiny progress bar}
\usage{
read_reports(report_files, report_names, cache_dir = "cache")
}
\arguments{
\item{report_files}{character vector with files}

\item{report_names}{character vector with names}
}
\value{
List of reports
}
\description{
Read report files, with Shiny progress bar
}

