% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datainput-read_reports.R
\name{read_reports}
\alias{read_reports}
\title{Read metagenomics classification results ('reports'), with Shiny progress bar}
\usage{
read_reports(report_files, report_names = NULL, cache_dir = NULL)
}
\arguments{
\item{report_files}{character vector with files, or a data.frame from read_sample_data, or a single directory path}

\item{report_names}{character vector with names}

\item{cache_dir}{cache directory path}
}
\value{
List of reports
}
\description{
Read metagenomics classification results ('reports'), with Shiny progress bar
}
\examples{
report_dir <- system.file("shinyapp", "example-data", "brain-biopsies", package="pavian")
reports <- read_reports(report_dir)
}
