% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datainput-read_sample_data.R
\name{read_sample_data}
\alias{read_sample_data}
\title{Read sample definitions from a directory}
\usage{
read_sample_data(my_dir, def_filename = "sample_data.csv", ext = c("report",
  "profile"), glob_files = FALSE)
}
\arguments{
\item{my_dir}{directory}

\item{def_filename}{filename}

\item{ext}{extension of report files}

\item{glob_files}{glob files?}
}
\value{
sample_data data.frame
}
\description{
Read sample definitions from a directory
}
