% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview-module.R
\name{reportOverviewModule}
\alias{reportOverviewModule}
\title{Shiny modules to display an overview of metagenomics reports}
\usage{
reportOverviewModule(input, output, session, sample_data, reports,
  datatable_opts = NULL)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiyn output object.}

\item{session}{Shiny session object.}

\item{sample_data}{Samples \code{data.frame}.}

\item{reports}{List of reports.}

\item{datatable_opts}{Additional datatable opts (mostly $class)}
}
\value{
Report overview module server functionality.
}
\description{
Shiny modules to display an overview of metagenomics reports
}
