% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runApp}
\alias{runApp}
\title{Run Centrifuger web interface}
\usage{
runApp(cache_dir = "cache", example_dir = system.file("shinyapp",
  "example-data", package = "pavian"), ...)
}
\arguments{
\item{cache_dir}{Directory to save temporary files.}

\item{...}{Additional arguments to \code{\link[shiny]{runApp}}, such as \code{host} and \code{port}.}
}
\description{
Run Centrifuger web interface
}

