% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runApp}
\alias{runApp}
\title{Run Centrifuger web interface}
\usage{
runApp(cache_dir = "cache")
}
\arguments{
\item{cache_dir}{Directory to save temporary files}
}
\description{
Run Centrifuger web interface
}

