% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runApp}
\alias{runApp}
\title{Run Pavian web interface}
\usage{
runApp(cache_dir = "cache", server_dir = Sys.glob("~"),
  server_access = FALSE, ...)
}
\arguments{
\item{cache_dir}{Directory to save temporary files.}

\item{server_dir}{Directory for sample files.}

\item{server_access}{Allow users to change server directory}

\item{...}{Additional arguments to \code{\link[shiny]{runApp}}, such as \code{host} and \code{port}.}
}
\description{
Run Pavian web interface
}

