% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-module.R
\name{sampleModule}
\alias{sampleModule}
\title{Server part for sample module}
\usage{
sampleModule(input, output, session, sample_data, reports, tax_data,
  clade_reads, taxon_reads, selected_sample = NULL,
  datatable_opts = NULL)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{sample_data}{Samples \code{data.frame}.}

\item{reports}{List of reports.}

\item{tax_data}{Taxonomy data.}

\item{clade_reads}{Clade reads.}

\item{taxon_reads}{Taxon reads.}

\item{selected_sample}{Pre-selected sample.}

\item{datatable_opts}{Additional datatable opts (mostly $class).}
}
\value{
Sample module server functionality
}
\description{
Server part for sample module
}
